new fees, force decay and typo

+++ src/interfaces/ILBRouter.sol
@@ -76,7 +76,7 @@ interface ILBRouter {
     function removeLiquidity(
         IERC20 tokenX,
         IERC20 tokenY,
-        uint8 binStep,
+        uint16 binStep,
         uint256 amountXMin,
         uint256 amountYMin,
         uint256[] memory ids,
@@ -87,7 +87,7 @@ interface ILBRouter {
 
     function removeLiquidityAVAX(
         IERC20 token,
-        uint8 binStep,
+        uint16 binStep,
         uint256 amountTokenMin,
         uint256 amountAVAXMin,
         uint256[] memory ids,
